moved global API_URL and API_PORT consts to api.js
