+++ vue-project/src/components/SimulationReport/simulation-report-generate.vue
@@ -31,8 +31,7 @@
 
 <script setup>
   import { ref, onMounted, watch, toRaw } from "vue";
-  import { API_URL, API_PORT } from "../../consts.js";
-  import { postSimulationReport } from "../api.js";
+  import { API_URL, API_PORT, postSimulationReport } from "../api.js";
   const {printJobs, workflows } = defineProps({
       printJobs: Array,
       workflows: Array,
