+++ vue-project/src/components/SubmitJobs.vue
@@ -35,7 +35,7 @@
 <script setup>
     import { ref, onMounted } from "vue";
     import { useRouter } from 'vue-router';
-    import { API_URL, API_PORT } from "../consts.js";
+    import { API_URL, API_PORT } from "./api.js";
 
     const router = useRouter();
     const routeTo = (where) => {
