+++ vue-project/src/components/Workflows.vue
@@ -51,8 +51,7 @@
 <script setup>
     import { ref, onMounted, toRaw } from "vue";
     import { useRouter } from 'vue-router';
-    import { API_URL, API_PORT } from "../consts.js";
-    import { getEntireCollection, formatLinearSteps, postWorkflow } from "./api.js";
+    import { API_URL, API_PORT, getEntireCollection, formatLinearSteps, postWorkflow } from "./api.js";
 
     const router = useRouter();
     const routeTo = (where) => {
