--- vue-project/src/consts.js
@@ -1,4 +1,0 @@
-// TODO: Vue supports environment variables stored in a .env file, 
-// which may be better than this method
-export const API_URL = "http://localhost"
-export const API_PORT = "5040"
